[diemdb-benchmark/executor-benchmark] refactor exec/diemdb benchmarks to store real txn data in db
Closes: #9247
